web-interface/django__backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
ansible-core==2.18.6
anthropic==0.54.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
billiard==4.2.1
black==25.1.0
blinker==1.9.0
Brotli==1.1.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
contourpy==1.3.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.9.1
cron-descriptor==1.4.5
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
dependency-injector==4.48.0
distlib==0.3.9
distro==1.9.0
Django==4.2.23
django-allauth==65.9.0
django-axes==8.0.0
django-bulk-update==2.2.0
django-cachalot==2.8.0
django-cache-memoize==0.2.1
django-cache-panel==0.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-csp==4.0
django-debreach==2.1.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==23.5
django-guardian==3.0.0
django-model-utils==5.0.0
django-querycount==0.8.3
django-ratelimit==4.1.0
django-redis==5.4.0
django-silk==5.4.0
django-sslserver==0.22
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docker==7.1.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
ecs-logging==2.2.0
elastic-apm==6.23.0
elastic-transport==8.17.1
elasticsearch==9.0.2
execnet==2.1.1
executing==2.2.0
factory_boy==3.3.3
faiss-cpu==1.11.0
Faker==37.4.0
filelock==3.18.0
flake8==7.2.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gevent==25.5.1
geventhttpclient==2.3.4
gitdb==4.0.12
GitPython==3.1.44
gprof2dot==2025.4.14
greenlet==3.2.3
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.1.4
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
identify==2.6.12
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
itsdangerous==2.2.0
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonfield==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.0.353
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langsmith==0.0.87
locust==2.37.10
locust-cloud==1.23.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.44.0
ncclient==0.6.19
netmiko==4.5.0
networkx==3.5
nodeenv==1.9.1
ntc_templates==7.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.88.0
orjson==3.10.18
packaging==23.2
pandas==2.3.0
paramiko==3.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
pre_commit==4.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.2.3
pyserial==3.5
pysnmp==7.1.20
pytest==7.4.4
pytest-asyncio==0.23.8
pytest-benchmark==5.0.1
pytest-cov==4.1.0
pytest-django==4.11.1
pytest-mock==3.14.1
pytest-profiling==1.8.1
pytest-xdist==3.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-memcached==1.62
python-nmap==0.7.1
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
redis==4.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
resolvelib==1.0.1
rich==14.0.0
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
scp==0.15.0
seaborn==0.13.2
sentence-transformers==2.7.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
textfsm==1.1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
vine==5.1.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.20.1
zope.event==5.0
zope.interface==7.2
